added positioning calculations for the Binary Tree only, Question 2

+++ App/src/Modules/App2/TreeNode2.java
@@ -1,8 +1,8 @@
 package App2;
 public class TreeNode2 {
     public char value;
-    public float x;
-    public float y;
+    public int x;
+    public int y;
 
     
 }
